few more changes

+++ src/dragen_align_pa/jobs/manage_dragen_pipeline.py
@@ -8,7 +8,7 @@ from cpg_utils.config import config_retrieve
 from cpg_utils.hail_batch import get_batch
 from hailtop.batch.job import PythonJob
 
-from src.dragen_align_pa.jobs import cancel_ica_pipeline_run, monitor_dragen_pipeline
+from src.dragen_align_pa.jobs import cancel_ica_pipeline_run, monitor_dragen_pipeline, run_align_genotype_with_dragen
 
 
 def initalise_management_job(sequencing_group: SequencingGroup, pipeline_id_file: str) -> PythonJob:
@@ -36,10 +36,6 @@ def _delete_pipeline_id_file(pipeline_id_file: str) -> None:
     subprocess.run(['gcloud', 'storage', 'rm', pipeline_id_file], check=True)  # noqa: S603, S607
 
 
-def _submit_ica_pipeline_job() -> None:
-    pass
-
-
 def manage_ica_pipeline(
     management_job: PythonJob,
     sequencing_group: SequencingGroup,
@@ -59,7 +55,7 @@ def manage_ica_pipeline(
     get_batch().write_output(management_output, output)
 
 
-def _run(sequencing_group: SequencingGroup, pipeline_id_file: str, api_root: str) -> dict[str, str]:
+def _run(sequencing_group: SequencingGroup, pipeline_id_file: str, api_root: str, output: str) -> dict[str, str]:
     # Get an existing pipeline ID
     with open(to_path(pipeline_id_file)) as pipeline_fid_handle:
         ica_pipeline_id: str = pipeline_fid_handle.read().rstrip()
@@ -71,17 +67,43 @@ def _run(sequencing_group: SequencingGroup, pipeline_id_file: str, api_root: str
         logging.info(f'Cancelling pipeline run: {ica_pipeline_id} for sequencing group {sequencing_group.name}')
         cancel_ica_pipeline_run.run(ica_pipeline_id=ica_pipeline_id, api_root=api_root)
         _delete_pipeline_id_file(pipeline_id_file=pipeline_id_file)
-    pipeline_status = monitor_dragen_pipeline.run(ica_pipeline_id=ica_pipeline_id, api_root=api_root)
-    if pipeline_status == 'SUCCEEDED':
-        logging.info(f'Pipeline run {ica_pipeline_id} has succeeded')
-        return {'pipeline': ica_pipeline_id, 'status': 'success'}
-    if pipeline_status in ['ABORTING', 'ABORTED']:
-        logging.info(f'The pipeline run {ica_pipeline_id} has been cancelled for sample {sequencing_group.name}.')
-        _delete_pipeline_id_file(pipeline_id_file=pipeline_id_file)
-        raise Exception(f'Pipeline run {ica_pipeline_id} has been cancelled.')
-    # Log failed ICA pipeline to a file somewhere
-    # Delete the pipeline ID file
-    _delete_pipeline_id_file(pipeline_id_file=pipeline_id_file)
-    raise Exception(f'The pipeline run {ica_pipeline_id} has failed, please check ICA for more info.')
 
-    _submit_ica_pipeline_job()
+    # Monitor an existing ICA pipeline run
+    if config_retrieve(['ica', 'management', 'monitor_previous'], False) and to_path(pipeline_id_file).exists():
+        pipeline_status = monitor_dragen_pipeline.run(ica_pipeline_id=ica_pipeline_id, api_root=api_root)
+        if pipeline_status == 'SUCCEEDED':
+            logging.info(f'Pipeline run {ica_pipeline_id} has succeeded')
+            return {'pipeline': ica_pipeline_id, 'status': 'success'}
+        if pipeline_status in ['ABORTING', 'ABORTED']:
+            logging.info(f'The pipeline run {ica_pipeline_id} has been cancelled for sample {sequencing_group.name}.')
+            _delete_pipeline_id_file(pipeline_id_file=pipeline_id_file)
+            raise Exception(f'Pipeline run {ica_pipeline_id} has been cancelled.')
+        # Log failed ICA pipeline to a file somewhere
+        # Delete the pipeline ID file
+        _delete_pipeline_id_file(pipeline_id_file=pipeline_id_file)
+        raise Exception(f'The pipeline run {ica_pipeline_id} has failed, please check ICA for more info.')
+
+    # Write ICA pipeline ID to file manually, rather than getting Hail Batch to do it?
+    # Submit a new ICA pipeline run
+    ica_pipeline_id = run_align_genotype_with_dragen.run(
+        ica_fids_path='x',
+        analysis_output_fid_path='x',
+        dragen_ht_id=config_retrieve(['ica', 'pipelines', 'dragen_ht_id']),
+        cram_reference_id=config_retrieve(
+            ['ica', 'cram_references', config_retrieve(['ica', 'cram_references', 'old_cram_reference'])]
+        ),
+        qc_cross_cont_vcf_id=config_retrieve(['ica', 'qc', 'cross_cont_vcf']),
+        qc_cov_region_1_id=config_retrieve(['ica', 'qc', 'coverage_region_1']),
+        qc_cov_region_2_id=config_retrieve(['ica', 'qc', 'coverage_region_2']),
+        dragen_pipeline_id=config_retrieve(['ica', 'pipelines', 'dragen_3_7_8']),
+        user_tags=config_retrieve(['ica', 'tags', 'user_tags']),
+        technical_tags=config_retrieve(['ica', 'tags', 'technical_tags']),
+        reference_tags=config_retrieve(['ica', 'tags', 'reference_tags']),
+        user_reference=sequencing_group.name,
+        api_root=api_root,
+        output_path=output,
+    )
+
+    monitor_dragen_pipeline.run(ica_pipeline_id=ica_pipeline_id, api_root=api_root)
+
+    return {'x': 'x'}
